+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Store entry point specification in context
 - Add context address size detection
 - Add binary data output in JSON export
 - Add binary data output in YAML export